Split boot protocols into modules
